feat: remove resample's internal data spec separation.
