+++ gwrdfa/resample/core/src/agreement.rs
@@ -19,29 +19,30 @@ use parabyzantine::agreement::{
 };
 use parabyzantine::{NoOp, NoOpData};
 pub use sampler::Sampler;
-pub use spec::ResampleAgreementSpec;
 pub use storage::ResampleAgreementStorage;
 pub use subcommittee::Subcommittee;
 
 /// A [ResampleAgreementBinding] is a binding for the [ResampleAgreement] protocol.
 ///
-/// It binds between the agreement data, the [ResampleAgreementSpec], and the [ResampleAgreementData].
+/// It binds between the agreement data and the [ResampleAgreementData].
 pub trait ResampleAgreementBinding: Sized
 where
 	<Self::ParabyzantineAgreementData as ParabyzantineAgreementData>::AgreementDraftBuffer:
 		ResampleAgreementStorage<
 			<Self::ParabyzantineAgreementData as ParabyzantineAgreementData>::AgreementEntity,
-			<Self::ResampleAgreementSpec as ResampleAgreementSpec<Self::ParabyzantineAgreementData>>::Index,
-			<Self::ResampleAgreementSpec as ResampleAgreementSpec<Self::ParabyzantineAgreementData>>::Subcommittee,
-			<Self::ResampleAgreementSpec as ResampleAgreementSpec<Self::ParabyzantineAgreementData>>::Value,
+			<Self::ResampleAgreementData as ResampleAgreementData<
+				Self::ParabyzantineAgreementData,
+			>>::Index,
+			<Self::ResampleAgreementData as ResampleAgreementData<
+				Self::ParabyzantineAgreementData,
+			>>::Subcommittee,
+			<Self::ResampleAgreementData as ResampleAgreementData<
+				Self::ParabyzantineAgreementData,
+			>>::Value,
 		>,
 {
 	type ParabyzantineAgreementData: ParabyzantineAgreementData;
-	type ResampleAgreementSpec: ResampleAgreementSpec<Self::ParabyzantineAgreementData>;
-	type ResampleAgreementData: ResampleAgreementData<
-		Self::ParabyzantineAgreementData,
-		Self::ResampleAgreementSpec,
-	>;
+	type ResampleAgreementData: ResampleAgreementData<Self::ParabyzantineAgreementData>;
 }
 
 /// [ResampleAgreement] wraps around the ResampleAgreement data indicated by the binding.
@@ -68,70 +69,6 @@ impl<Binding: ResampleAgreementBinding> ResampleAgreement<Binding>
 	}
 }
 
-impl<Binding: ResampleAgreementBinding>
-	ResampleAgreementData<
-		Binding::ParabyzantineAgreementData,
-		Binding::ResampleAgreementSpec,
-	> for ResampleAgreement<Binding>
-{
-	/// A [ResampleAgreement] data must be able to provide a [CertificateSet]
-	fn certificate_set(
-		&self,
-	) -> &<Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::CertificateSet {
-		self.data().certificate_set()
-	}
-
-	/// A [ResampleAgreement] data must be able to provide a mutable [CertificateSet]
-	fn certificate_set_mut(
-		&mut self,
-	) -> &mut <Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::CertificateSet {
-		self.data_mut().certificate_set_mut()
-	}
-
-	/// A [ResampleAgreement] data must be able to provide a [Sampler]
-	fn sampler(
-		&self,
-	) -> &<Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::Sampler {
-		self.data().sampler()
-	}
-
-	/// A [ResampleAgreement] data must be able to provide a mutable [Sampler]
-	fn sampler_mut(
-		&mut self,
-	) -> &mut <Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::Sampler {
-		self.data_mut().sampler_mut()
-	}
-
-	/// A [ResampleAgreement] data must be able to provide a [IndexSubcommitteeAgreementQuery]
-	fn index_subcommittee_agreement_query_plan(
-		&mut self,
-	) -> <Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::IndexSubcommitteeAgreementQueryPlan {
-		self.data_mut().index_subcommittee_agreement_query_plan()
-	}
-
-	/// A [ResampleAgreement] data must be able to provide a [CertificateQuery]
-	fn certificate_query_plan(
-		&mut self,
-		index: &<Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-			Binding::ParabyzantineAgreementData,
-		>>::Index,
-	) -> <Binding::ResampleAgreementSpec as ResampleAgreementSpec<
-		Binding::ParabyzantineAgreementData,
-	>>::CertificateQueryPlan {
-		self.data_mut().certificate_query_plan(index)
-	}
-}
-
 impl<Binding: ResampleAgreementBinding> ParabyzantineAgreement<Binding::ParabyzantineAgreementData>
 	for ResampleAgreement<Binding>
 {
@@ -140,18 +77,18 @@ impl<Binding: ResampleAgreementBinding> ParabyzantineAgreement<Binding::Parabyza
 		agreement_world: &mut AgreementWorld<Binding::ParabyzantineAgreementData>,
 	) {
 		// over all the index subcommittee agreements
-		let index_query = self.index_subcommittee_agreement_query_plan();
+		let index_query = self.data_mut().index_subcommittee_agreement_query_plan();
 		for (_agreement_entity, (index, subcommittee)) in
 			agreement_world.agreement_facts.query(index_query)
 		{
 			// Insert all of the certificates for this index into the certificate set
-			let certificate_query_plan = self.certificate_query_plan(index);
+			let certificate_query_plan = self.data_mut().certificate_query_plan(index);
 			// insert all of the certificates for this index into the certificate set
 			for (_certificate_entity, (index, value, subcommittee)) in
 				agreement_world.certificate_facts.query(certificate_query_plan)
 			{
 				// This is just for moving the certificate into the certificate set.
-				self.certificate_set_mut().insert(
+				self.data_mut().certificate_set_mut().insert(
 					index.clone(),
 					value.clone(),
 					subcommittee.clone(),
@@ -159,15 +96,16 @@ impl<Binding: ResampleAgreementBinding> ParabyzantineAgreement<Binding::Parabyza
 			}
 
 			// check the subcommittee condition
-			let subcommittee_condition = subcommittee
-				.condition(self.certificate_set().partial_subcommittees_for_index(index));
+			let subcommittee_condition = {
+				let certificate_set = self.data().certificate_set();
+				subcommittee.condition(certificate_set.partial_subcommittees_for_index(index))
+			};
 
 			// elect the next subcommittee from the condition
-			let next_subcommittee = self.sampler_mut().elect_subcommittee_from_condition(
-				index,
-				subcommittee,
-				&subcommittee_condition,
-			);
+			let next_subcommittee = self
+				.data_mut()
+				.sampler_mut()
+				.elect_subcommittee_from_condition(index, subcommittee, &subcommittee_condition);
 
 			if let Some((next_index, next_subcommittee)) = next_subcommittee {
 				// insert the next subcommittee into the agreement world
@@ -199,10 +137,7 @@ impl<Binding: ResampleAgreementBinding> ResampleAgreement<Binding> {
 	/// A direct implementation of resampling on an agreement world.
 	///
 	/// This is most useful for experimenting and testing.
-	pub fn resample_agreement(
-		&mut self,
-		agreement_data: &mut Binding::ParabyzantineAgreementData,
-	) {
+	pub fn resample_agreement(&mut self, agreement_data: &mut Binding::ParabyzantineAgreementData) {
 		let mut agreement_world = agreement_data.parabyzantine_agreement_world();
 
 		self.update_parabyzantine_agreement(&mut agreement_world);
@@ -214,7 +149,6 @@ impl<Binding: ResampleAgreementBinding> ResampleAgreement<Binding> {
 /// A [ResampleAgreementBinding] for the [NoOp] struct.
 impl ResampleAgreementBinding for NoOp {
 	type ParabyzantineAgreementData = NoOpData;
-	type ResampleAgreementSpec = NoOp;
 	type ResampleAgreementData = NoOpData;
 }
 
@@ -222,7 +156,6 @@ impl ResampleAgreementBinding for NoOp {
 mod tests {
 	use super::*;
 	use crate::agreement::data::test::TestResampleAgreementData;
-	use crate::agreement::spec::test::TestResampleAgreementSpec;
 	use crate::agreement::subcommittee::test::TestSubcommittee;
 	use crate::agreement::test_util::container::TestResampleAgreementContainer;
 	use crate::agreement::test_util::container::TestResampleCertificateContainer;
@@ -231,8 +164,8 @@ mod tests {
 	use crate::task::ResampleTask;
 	use gwrdfa_container::Component;
 	use parabyzantine::{
-		agreement::Agreement, task::Task, AgreementAction, AgreementHandler, Data,
-		Parabyzantine, Spec, TaskAction, TaskHandler,
+		agreement::Agreement, task::Task, AgreementAction, AgreementHandler, Data, Parabyzantine,
+		Spec, TaskAction, TaskHandler,
 	};
 	use std::collections::BTreeSet;
 	use std::vec;
@@ -240,7 +173,6 @@ mod tests {
 	impl ResampleAgreementBinding for TestResampleAgreementData<u32, u32, TestSubcommittee<u32>> {
 		type ParabyzantineAgreementData =
 			TestResampleParabyzantineData<u32, u32, TestSubcommittee<u32>>;
-		type ResampleAgreementSpec = TestResampleAgreementSpec<u32, u32, TestSubcommittee<u32>>;
 		type ResampleAgreementData = TestResampleAgreementData<u32, u32, TestSubcommittee<u32>>;
 	}
