+++ gwrdfa/resample/core/src/agreement/spec.rs
@@ -1,130 +1,1 @@
-use super::{CertificateSet, ResampleAgreementStorage, Sampler, Subcommittee};
-use parabyzantine::NoOp;
-use parabyzantine::{
-	agreement::ParabyzantineAgreementData,
-	buffer::query::{QueryPlanlike, Querylike},
-	NoOpData,
-};
-
-/// A [ResampleAgreementSpec] is a specification for ResampleAgreement consensus.
-pub trait ResampleAgreementSpec<Data: ParabyzantineAgreementData>: Sized
-where
-	Data::AgreementDraftBuffer:
-		ResampleAgreementStorage<
-			Data::AgreementEntity,
-			Self::Index,
-			Self::Subcommittee,
-			Self::Value,
-		>,
-{
-	/// The type of the index.
-	/// The index must be clonable in order to be able to insert index subcommittee agreement facts.
-	type Index: Clone + Eq;
-
-	/// The type of the value.
-	/// The value must be clonable in order to be able to insert value agreement facts.
-	type Value: Clone + Eq + 'static;
-
-	/// The type of the subcommittee.
-	/// The subcommittee must be clonable in order to be able to insert subcommittee agreement facts.
-	type Subcommittee: Subcommittee<Self::Value> + Clone;
-
-	/// The query for the index subcommittee agreement.
-	type IndexSubcommitteeAgreementQuery<'a>: Querylike<
-		Data::AgreementEntity,
-		(&'a Self::Index, &'a Self::Subcommittee),
-	>
-	where
-		Self::Index: 'a,
-		Self::Subcommittee: 'a;
-
-	/// The query plan for the index subcommittee agreement.
-	type IndexSubcommitteeAgreementQueryPlan: for<'a> QueryPlanlike<
-		Data::AgreementEntity,
-		&'a Data::AgreementBuffer,
-		(&'a Self::Index, &'a Self::Subcommittee),
-		Self::IndexSubcommitteeAgreementQuery<'a>,
-	>;
-
-	/// The query for the certificate.
-	type CertificateQuery<'a>: Querylike<
-		Data::CertificateEntity,
-		(&'a Self::Index, &'a Self::Value, &'a Self::Subcommittee),
-	>
-	where
-		Self::Index: 'a,
-		Self::Value: 'a,
-		Self::Subcommittee: 'a;
-
-	/// The query plan for the certificate.
-	type CertificateQueryPlan: for<'a> QueryPlanlike<
-		Data::CertificateEntity,
-		&'a Data::CertificateBuffer,
-		(&'a Self::Index, &'a Self::Value, &'a Self::Subcommittee),
-		Self::CertificateQuery<'a>,
-	>;
-
-	/// The type of the certificate set.
-	type CertificateSet: CertificateSet<Self::Index, Self::Value, Self::Subcommittee>;
-
-	/// The type of the sampler.
-	type Sampler: Sampler<Self::Index, Self::Value, Self::Subcommittee>;
-}
-
-impl ResampleAgreementSpec<NoOpData> for NoOp
-where
-	NoOp: ResampleAgreementStorage<NoOp, NoOp, NoOp, NoOp>,
-{
-	type Index = NoOp;
-	type Value = NoOp;
-	type Subcommittee = NoOp;
-	type IndexSubcommitteeAgreementQuery<'a> = NoOp;
-	type IndexSubcommitteeAgreementQueryPlan = NoOp;
-	type CertificateQuery<'a> = NoOp;
-	type CertificateQueryPlan = NoOp;
-	type CertificateSet = NoOp;
-	type Sampler = NoOp;
-}
-
-#[cfg(test)]
-pub mod test {
-	use super::*;
-	use crate::agreement::certificate::test::TestCertificateSet;
-	use crate::agreement::sampler::test::TestSampler;
-	use crate::agreement::test_util::container::{
-		TestResampleAgreementContainer, TestResampleCertificateContainer,
-		TestResampleParabyzantineData,
-	};
-	use crate::agreement::test_util::TextIndexabled;
-	use crate::agreement::test_util::{Index, Sub, Value};
-	use core::marker::PhantomData;
-	use gwrdfa_container::query::matching_tuple::{MatchingTuple, MatchingTupleQuery};
-	use std::hash::Hash;
-
-	pub struct TestResampleAgreementSpec<Index: Eq, Value: Eq + 'static, Sub: Subcommittee<Value>> {
-		__marker: PhantomData<(Index, Value, Sub)>,
-	}
-
-	impl<
-			I: Eq + Clone + Hash + TextIndexabled + 'static,
-			V: Eq + Clone + Hash + 'static,
-			S: Subcommittee<V> + Clone + Hash + 'static,
-		> ResampleAgreementSpec<TestResampleParabyzantineData<I, V, S>>
-		for TestResampleAgreementSpec<I, V, S>
-	{
-		type Index = Index<I>;
-		type Value = Value<V>;
-		type Subcommittee = Sub<S>;
-		type IndexSubcommitteeAgreementQuery<'a> =
-			MatchingTupleQuery<'a, TestResampleAgreementContainer<I, V, S>, (Index<I>, Sub<S>)>;
-		type IndexSubcommitteeAgreementQueryPlan = MatchingTuple<(Index<I>, Sub<S>)>;
-		type CertificateQuery<'a> = MatchingTupleQuery<
-			'a,
-			TestResampleCertificateContainer<I, V, S>,
-			(Index<I>, Value<V>, Sub<S>),
-		>;
-		type CertificateQueryPlan = MatchingTuple<(Index<I>, Value<V>, Sub<S>)>;
-		type CertificateSet = TestCertificateSet<Index<I>, Value<V>, Sub<S>>;
-		type Sampler = TestSampler;
-	}
-}
+pub use super::data::ResampleAgreementData as ResampleAgreementSpec;
