+++ gwrdfa/resample/core/src/task/spec.rs
@@ -1,71 +1,1 @@
-use super::{IndexTaskSubcommitteeAgreement, ResampleTasker, TaskSubcommittee};
-use parabyzantine::{
-	buffer::query::{QueryPlanlike, Querylike},
-	task::ParabyzantineTaskData,
-	NoOp,
-	NoOpData,
-};
-
-pub trait ResampleTaskSpec<Data: ParabyzantineTaskData>: Sized {
-	/// The type of the index.
-	type Index: Eq;
-
-	/// The type of the sender of a task.
-	type Sender: Eq;
-
-	/// The type of the value of a task.
-	type Value: Eq + 'static;
-
-	/// The type of the task subcommittee.
-	type TaskSubcommittee: TaskSubcommittee<Self::Value, Self::Sender>;
-
-	/// The type queried for indicating agreement on an index.
-	type IndexTaskSubcommitteeAgreementQueryData<'a>;
-
-	/// The query for the index subcommittee agreement.
-	type IndexTaskSubcommitteeAgreementQuery<'a>: Querylike<
-		Data::AgreementEntity,
-		Self::IndexTaskSubcommitteeAgreementQueryData<'a>,
-	>;
-
-	/// The query plan for the index subcommittee agreement.
-	type IndexTaskSubcommitteeAgreementQueryPlan: for<'a> QueryPlanlike<
-		Data::AgreementEntity,
-		&'a Data::AgreementBuffer,
-		Self::IndexTaskSubcommitteeAgreementQueryData<'a>,
-		Self::IndexTaskSubcommitteeAgreementQuery<'a>,
-	>;
-
-	/// The type of the index subcommittee agreement.
-	type IndexTaskSubcommitteeAgreement: IndexTaskSubcommitteeAgreement<
-			Self::Index,
-			Self::Value,
-			Self::Sender,
-			Self::TaskSubcommittee,
-		> + for<'a> From<(
-			Data::AgreementEntity,
-			Self::IndexTaskSubcommitteeAgreementQueryData<'a>,
-		)>;
-
-	/// The tasker
-	type ResampleTasker: ResampleTasker<
-		Self::Index,
-		Self::Sender,
-		Self::Value,
-		Self::TaskSubcommittee,
-		Self::IndexTaskSubcommitteeAgreement,
-		Data,
-	>;
-}
-
-impl ResampleTaskSpec<NoOpData> for NoOp {
-	type Index = NoOp;
-	type Sender = NoOp;
-	type Value = NoOp;
-	type TaskSubcommittee = NoOp;
-	type IndexTaskSubcommitteeAgreementQueryData<'a> = NoOp;
-	type IndexTaskSubcommitteeAgreementQuery<'a> = NoOp;
-	type IndexTaskSubcommitteeAgreementQueryPlan = NoOp;
-	type IndexTaskSubcommitteeAgreement = NoOp;
-	type ResampleTasker = NoOp;
-}
+pub use super::data::ResampleTaskData as ResampleTaskSpec;
